requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cycler==0.10.0
dataclasses==0.6
future==0.18.2
idna==2.10
img2vec-pytorch==0.2.5
joblib==0.17.0
kiwisolver==1.3.1
kneed==0.7.0
lxml==4.6.1
matplotlib==3.3.2
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.4
pathspec==0.8.0
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
regex==2020.10.28
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
webdriver-manager==3.2.2
